add models.py and moved guest class out of app.py
